requirements.txt
pydub
openai-whisper
numpy==1.22